Upgrade django-taggit to version with migrations fix.
Previous version had an issue with the 0001-initial migrations file,
such that a 0002-auto migration would be created in the taggit
site-packages. This caused issues between machines, since the auto
migration filename includes a timestamp, and other migrations within
my project refer to it as a dependency.

So, I upgraded to the next version of taggit, which fixed the issue.